\name{monNouveauPackage-package}
\alias{monNouveauPackage-package}
\alias{monNouveauPackage}
\docType{package}
\title{
\packageTitle{monNouveauPackage}
}
\description{
\packageDescription{monNouveauPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{monNouveauPackage}
\packageIndices{monNouveauPackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{monNouveauPackage}

Maintainer: \packageMaintainer{monNouveauPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
