% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multplus.R
\name{mult}
\alias{mult}
\title{multiplier 2 nombres}
\usage{
mult(x, y)
}
\arguments{
\item{x}{Un nombre}

\item{y}{Un nombre}
}
\value{
Le produit de \code{x} et \code{y}
}
\description{
multiplier 2 nombres
}
\examples{
\dontrun{
mult(2, 3)
mult(10, 2)
}
}
