% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multplus.R
\name{plus}
\alias{plus}
\title{sommer 2 nombres}
\usage{
plus(x, y)
}
\arguments{
\item{x}{Un nombre}

\item{y}{Un nombre}
}
\value{
La somme de \code{x} et \code{y}
}
\description{
sommer 2 nombres
}
\examples{
\dontrun{
plus(2, 3)
plus(10, 20)
}
}
