% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeRand.R
\name{treeFunKeyrus}
\alias{treeFunKeyrus}
\title{Arbre decision sur une matrice aleatoire}
\usage{
treeFunKeyrus(nbLig, nbCol)
}
\arguments{
\item{nbLig}{nombre de ligne}

\item{nbCol}{nombre de colonne}
}
\value{
Le resume de modele arbre decisin sur \code{nbLig} ligne et \code{nbCol} colonne
}
\description{
Arbre decision sur une matrice aleatoire
}
\examples{
treeFunKeyrus(10, 2)
mult(100, 5)
}
